requirements.txxt.txt
Flask==2.3.3
mercadopago==2.0.0
flask-cors==4.0.0
python-dotenv==1.0.0